requirements.txt
django~=5.0.4
django-cors-headers~=4.3.1
djangorestframework~=3.15.1
drf-yasg~=1.21.7
gunicorn~=21.2.0
pytest~=8.1.1
pytest-django~=4.8.0
python-dotenv==1.0.0